+++ utils/blsaes/blsaes_test.go
@@ -1,11 +1,11 @@
-package aes_test
+package blsaes_test
 
 import (
 	"reflect"
 	"testing"
 
 	"github.com/olympus-protocol/ogen/bls"
-	"github.com/olympus-protocol/ogen/utils/aes"
+	"github.com/olympus-protocol/ogen/utils/blsaes"
 )
 
 var encKey = []byte("test")
@@ -13,11 +13,11 @@ var encKey = []byte("test")
 func Test_EncryptDecrypt(t *testing.T) {
 	rand := bls.RandKey()
 	keyBytes := rand.Marshal()
-	nonce, salt, cipher, err := aes.Encrypt(keyBytes, encKey)
+	nonce, salt, cipher, err := blsaes.Encrypt(keyBytes, encKey)
 	if err != nil {
 		t.Fatal(err)
 	}
-	key, err := aes.Decrypt(cipher, nonce, encKey, salt[:])
+	key, err := blsaes.Decrypt(cipher, nonce, encKey, salt[:])
 	if err != nil {
 		t.Fatal(err)
 	}
